backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
Werkzeug==3.1.3
flask-cors==4.0.0
psycopg2-binary==2.9.9
tensorflow==2.15.0
tensorflow-hub==0.15.0
numpy==1.26.4
